requirements.txt
asgiref==3.8.1
awesome-slugify==1.6.5
cffi==1.17.1
Django==5.1.7
django-etc==1.4.0
django-hitcount==1.3.5
django-js-asset==3.1.2
django-prose-editor==0.10.4
pillow==11.1.0
pycparser==2.22
python-slugify==8.0.4
pytz==2025.2
regex==2024.11.6
sorl-thumbnail==12.11.0
sqlparse==0.5.3
text-unidecode==1.3
tzdata==2025.1
Unidecode==0.4.21
urlquote==2.1.0
